8-1掲示板の作成/requirements.txt
Flask==2.0.2
psycopg2-binary == 2.9.6
Flask-HTTPAuth == 4.8.0
gunicorn
